Split App.tsx into modules, with the interface byte-identical (#186)
App.tsx held the whole web UI in 2413 lines, 1385 of them inside App() itself.
It goes to 1393, with the three-language string table, the KaTeX renderer, the
seven leaf components, the three status screens and the settings overlay moved
into src/lib/ and src/components/.

This moves code and changes nothing else, because the visual redesign lands
next and a diff that does both cannot be reviewed. The property that makes that
claim checkable: every className string in App.tsx at the previous commit is
still present byte for byte across the resulting files, and none was added.

The sidebar, chat column and composer stay in App.tsx on purpose. They read
enough of App()'s state that extracting them now would produce thirty-prop
components, which is worse than the long file and would get in the redesign's
way.

Two components take the string table itself rather than resolved strings, so
they are typed against it (`Strings` in lib/i18n.ts). Record<string, string>
there would let a mistyped key compile and render `undefined`; the status
screens take resolved strings instead and need no table at all.
