+++ rag/web/frontend/src/App.tsx
@@ -1,358 +1,35 @@
 import React, { useState, useRef, useEffect, useCallback } from 'react';
 import {
   Send, FileText, Database, Ollama,
-  Search, Layers, FileUp, Menu, X,
+  Search, Layers, Menu, X,
   RefreshCw, Loader2, AlertCircle, CheckCircle2, Trash2,
-  ChevronDown, Copy, Check, Languages, Eye,
-  Power, Sun, Moon, BookOpen
+  Copy, Check, Eye,
+  Sun, Moon, BookOpen
 } from './lib/icons';
 import { getStoredTheme, setTheme, type Theme } from './lib/theme';
 import { motion, AnimatePresence } from 'motion/react';
-import katex from 'katex';
-import 'katex/dist/katex.min.css';
-
-// =============================================================================
-// Types
-// =============================================================================
-
-// 'study' is a front-end-only mode: the backend's /api/mode knows chat and rag,
-// and a study turn is a POST to /api/study rather than a mode the server holds.
-type Mode = 'chat' | 'rag' | 'study';
-type StudyKind = 'summary' | 'outline' | 'quiz';
-
-interface OutlineNode { title: string; children: OutlineNode[]; }
-interface StudyArtifact {
-  source_document?: string;
-  sections?: { heading: string; body: string; source_pages: number[] }[];
-  nodes?: OutlineNode[];
-  questions?: { prompt: string; options: string[]; correct_index: number; source_pages: number[] }[];
-}
-
-interface Citation {
-  document: string;
-  pages: number[];
-  best_page?: number;
-}
-
-interface Message {
-  id: string;
-  role: 'user' | 'assistant' | 'system';
-  content: string;
-  mode: Mode;
-  citations?: Citation[];
-  metrics?: { searchTime: string; chunks: number };
-  isStreaming?: boolean;
-  isError?: boolean;
-  // A study turn answers with structure rather than prose, so it is carried
-  // as data and rendered by the bubble instead of being flattened to text.
-  artifact?: StudyArtifact;
-  artifactKind?: StudyKind;
-}
-
-interface PipelineSettings {
-  contextualRetrieval: boolean;
-  queryDecomposition: boolean;
-  hybridSearch: boolean;
-  imageIndexing: boolean;
-  imageDescription: boolean;
-  reranker: boolean;
-  expandContext: boolean;
-  optimizeContext: boolean;
-  recompSynthesis: boolean;
-}
-
-interface IndexingProgress {
-  file: string;
-  file_index: number;
-  total_files: number;
-}
-
-interface OllamaModel {
-  name: string;
-  size?: number;
-  family?: string;
-  parameter_size?: string;
-  capabilities?: string[];
-  embedding?: boolean;
-  vision?: boolean;
-}
-
-type ModelRole = 'rag' | 'chat' | 'contextual' | 'recomp';
-type ModelRoles = Record<ModelRole, string>;
-
-interface VectorStore {
-  name: string;
-  label: string;
-  docs_folder: string;
-  pdf_count: number;
-  indexed: boolean;
-  active: boolean;
-  fragments: number | null;
-}
-
-// =============================================================================
-// i18n
-// =============================================================================
-
-type Lang = 'es' | 'en' | 'ca';
-
-const STRINGS = {
-  es: {
-    tabDocs: 'Documentos', tabPipeline: 'Pipeline RAG',
-    corpusError: 'No se pudo cambiar el almacén.',
-    corpusConflict: 'Hay indexación en curso; espera un momento.',
-    corpusConnError: '✗ Error de conexión al cambiar el corpus.',
-    collection: 'Colección ({n} docs)', noDocs: 'No hay documentos indexados',
-    deleteDoc: 'Eliminar documento', viewPdf: 'Ver PDF',
-    indexingFile: 'Indexando {file}', reindexingStatus: 'Re-indexación en curso...',
-    section1: '1. Indexación',
-    labelContextual: 'Recuperación contextual', descContextual: 'Enriquece fragmentos con LLM',
-    section2: '2. Recuperación',
-    labelHybrid: 'Búsqueda híbrida', descHybrid: 'Semántica + BM25',
-    labelQueryDecomp: 'Descomposición de consultas', descQueryDecomp: 'Subconsultas con LLM auxiliar',
-    labelImageIndex: 'Indexado de imágenes', descImageIndex: 'Recuperación visual directa',
-    labelImageDesc: 'Descripción de figuras', descImageDesc: 'Texto descriptivo por visión (requiere reindexado y modelo con visión)',
-    section3: '3. Ranking y contexto',
-    labelReranker: 'Reordenador cross-encoder', descReranker: 'Reordenamiento de precisión',
-    labelExpandContext: 'Expandir contexto', descExpandContext: 'Añade fragmentos adyacentes',
-    labelOptimizeContext: 'Optimizar contexto', descOptimizeContext: 'Limpia artefactos PDF',
-    labelRecomp: 'Síntesis RECOMP', descRecomp: 'Sintetiza contexto con LLM',
-    section4: '4. Reindexación',
-    reindexHint: 'Ajusta las opciones arriba y reindexa para aplicarlas. Opcionalmente añade PDFs nuevos.',
-    addPdfs: 'Añadir PDFs', remove: 'Quitar', reindexBtn: 'Reindexar',
-    fragments: '{n} fragmentos', ollamaStatus: 'Ollama Local',
-    clearChat: 'Limpiar chat', youLabel: 'Tú',
-    themeLight: 'Modo claro', themeDark: 'Modo oscuro', close: 'Cerrar',
-    sources: '{n} fuentes', copyMsg: 'Copiar mensaje',
-    placeholderRag: 'Pregunta sobre tus documentos…', placeholderChat: 'Escribe un mensaje…',
-    footerMode: 'Modo documento', footerModeChat: 'Modo conversación',
-    indexingTitle: 'Se están indexando los documentos',
-    indexingHint: 'Puede tardar unos minutos dependiendo de tu hardware.',
-    processing: 'Procesando:', fileUnit: 'archivo', fileUnitPlural: 'archivos',
-    autoRefresh: 'La página se actualizará automáticamente al terminar.',
-    retry: 'Reintentar', connErrorTitle: 'Error de conexión',
-    connecting: 'Conectando con MonkeyGrab…',
-    reindexError: '✗ Error: {error}', reindexConnError: '✗ Error de conexión.',
-    confirmDelete: '¿Eliminar "{name}"? Se borrará del índice y del disco.',
-    docDeleted: '✓ Documento **{name}** eliminado.',
-    deleteError: '✗ Error al eliminar: {error}',
-    deleteConnError: '✗ Error de conexión al eliminar.',
-    unknownError: 'Error desconocido', queryError: 'Error en la consulta',
-    modelError: 'Error en la respuesta del modelo',
-    connClosed: 'La conexión se cerró antes de completar la respuesta.',
-    settingsSaveError: 'No se pudo guardar el ajuste.',
-    noServer: 'No se pudo conectar con el servidor. ¿Está Flask ejecutándose?',
-    retryError: 'Error al reintentar',
-    indexingFailed: 'La indexación no pudo completarse.',
-    noResults: 'No se encontró información relevante en los documentos.',
-    tabModels: 'Modelos',
-    modeStudy: 'ESTUDIAR',
-    studyKindSummary: 'Resumen', studyKindOutline: 'Esquema', studyKindQuiz: 'Cuestionario',
-    studyPlaceholder: 'Escribe el nombre de un documento...',
-    studyHint: 'Elige qué generar y escribe el documento. Tab para autocompletar.',
-    studyNoMatch: 'No encuentro ningún documento que coincida con «{query}».',
-    studyAnswer: 'Respuesta', studyShowAnswersInline: 'Ver respuestas',
-    studyEmptyTitle: 'Estudia tus documentos',
-    studyEmptyBody: 'Un resumen, un esquema o un cuestionario de cualquier PDF indexado.',
-    studyPages: 'Páginas', studyPagesShort: 'páginas',
-    studyMalformed: 'El modelo no respetó el formato. Prueba otra vez, o con otro modelo.',
-    studyFailed: 'No se pudo generar: {error}',
-    storesLabel: 'Almacén vectorial',
-    storeNotIndexed: 'sin indexar',
-    storeEmptyHint: 'Almacén vacío. Sube PDFs y reindexa para activarlo.',
-    fingerprintStaleWarning: 'El índice guardado no coincide con la configuración activa, pero la app sigue funcionando con él. Reindexa cuando quieras: puede tardar una hora o más.',
-    modelsRoles: 'Roles de modelo',
-    ollamaTitle: 'Servidor Ollama',
-    ollamaOnline: 'En ejecución',
-    ollamaOffline: 'Detenido',
-    ollamaStartBtn: 'Arrancar Ollama',
-    ollamaStarting: 'Arrancando…',
-    ollamaStartFailed: 'No se pudo arrancar Ollama.',
-    ollamaNotFound: 'No se encontró el ejecutable de Ollama en el PATH.',
-    refreshModels: 'Actualizar lista',
-    noModels: 'No hay modelos instalados. Descárgalos con «ollama pull».',
-    roleRag: 'Generador RAG', descRoleRag: 'Respuesta final en modo documento',
-    roleChat: 'Subconsultas', descRoleChat: 'Conversación y descomposición de consultas',
-    roleContextual: 'Recuperación contextual', descRoleContextual: 'Enriquece fragmentos al indexar',
-    roleRecomp: 'Síntesis RECOMP', descRoleRecomp: 'Resume el contexto antes de generar',
-    modelSaveError: 'No se pudo cambiar el modelo.',
-  },
-  en: {
-    tabDocs: 'Documents', tabPipeline: 'RAG Pipeline',
-    corpusError: 'Could not switch store.',
-    corpusConflict: 'Indexing in progress; please wait.',
-    corpusConnError: '✗ Connection error while switching corpus.',
-    collection: 'Collection ({n} docs)', noDocs: 'No documents indexed',
-    deleteDoc: 'Delete document', viewPdf: 'View PDF',
-    indexingFile: 'Indexing {file}', reindexingStatus: 'Re-indexing in progress...',
-    section1: '1. Indexing',
-    labelContextual: 'Contextual Retrieval', descContextual: 'Enrich chunks with LLM',
-    section2: '2. Retrieval',
-    labelHybrid: 'Hybrid Search', descHybrid: 'Semantic + BM25',
-    labelQueryDecomp: 'Query Decomposition', descQueryDecomp: 'Sub-queries via auxiliary LLM',
-    labelImageIndex: 'Image indexing', descImageIndex: 'Direct visual retrieval',
-    labelImageDesc: 'Figure descriptions', descImageDesc: 'Vision-written figure text (needs reindex and a vision model)',
-    section3: '3. Ranking & Context',
-    labelReranker: 'Cross-Encoder Reranker', descReranker: 'Precision reordering',
-    labelExpandContext: 'Expand Context', descExpandContext: 'Add adjacent chunks',
-    labelOptimizeContext: 'Optimize Context', descOptimizeContext: 'Clean PDF artifacts',
-    labelRecomp: 'RECOMP Synthesis', descRecomp: 'Synthesize context with LLM',
-    section4: '4. Re-index',
-    reindexHint: 'Adjust settings above and re-index to apply. Optionally add new PDFs.',
-    addPdfs: 'Add PDFs', remove: 'Remove', reindexBtn: 'Re-index',
-    fragments: '{n} fragments', ollamaStatus: 'Ollama Local',
-    clearChat: 'Clear chat', youLabel: 'You',
-    themeLight: 'Light mode', themeDark: 'Dark mode', close: 'Close',
-    sources: '{n} sources', copyMsg: 'Copy message',
-    placeholderRag: 'Ask about your documents…', placeholderChat: 'Type a message…',
-    footerMode: 'Document mode', footerModeChat: 'Conversation mode',
-    indexingTitle: 'Indexing documents',
-    indexingHint: 'This may take a few minutes depending on your hardware.',
-    processing: 'Processing:', fileUnit: 'file', fileUnitPlural: 'files',
-    autoRefresh: 'The page will refresh automatically when done.',
-    retry: 'Retry', connErrorTitle: 'Connection error',
-    connecting: 'Connecting to MonkeyGrab…',
-    reindexError: '✗ Error: {error}', reindexConnError: '✗ Connection error.',
-    confirmDelete: 'Delete "{name}"? It will be removed from the index and disk.',
-    docDeleted: '✓ Document **{name}** deleted.',
-    deleteError: '✗ Error deleting: {error}',
-    deleteConnError: '✗ Connection error while deleting.',
-    unknownError: 'Unknown error', queryError: 'Query error',
-    modelError: 'Error in model response',
-    connClosed: 'The connection closed before the response was complete.',
-    settingsSaveError: 'Could not save setting.',
-    noServer: 'Could not connect to the server. Is Flask running?',
-    retryError: 'Retry failed',
-    indexingFailed: 'Indexing could not be completed.',
-    noResults: 'No relevant information found in the documents.',
-    tabModels: 'Models',
-    modeStudy: 'STUDY',
-    studyKindSummary: 'Summary', studyKindOutline: 'Outline', studyKindQuiz: 'Quiz',
-    studyPlaceholder: 'Type a document name...',
-    studyHint: 'Pick what to generate and type the document. Tab to autocomplete.',
-    studyNoMatch: 'No document matches "{query}".',
-    studyAnswer: 'Answer', studyShowAnswersInline: 'Show answers',
-    studyEmptyTitle: 'Study your documents',
-    studyEmptyBody: 'A summary, an outline or a quiz from any indexed PDF.',
-    studyPages: 'Pages', studyPagesShort: 'pages',
-    studyMalformed: 'The model did not follow the format. Try again, or another model.',
-    studyFailed: 'Could not generate: {error}',
-    storesLabel: 'Vector store',
-    storeNotIndexed: 'not indexed',
-    storeEmptyHint: 'Empty store. Upload PDFs and re-index to activate it.',
-    fingerprintStaleWarning: 'The saved index no longer matches the active configuration, but the app keeps working with it. Re-index whenever you like: it can take an hour or more.',
-    modelsRoles: 'Model roles',
-    ollamaTitle: 'Ollama server',
-    ollamaOnline: 'Running',
-    ollamaOffline: 'Stopped',
-    ollamaStartBtn: 'Start Ollama',
-    ollamaStarting: 'Starting…',
-    ollamaStartFailed: 'Could not start Ollama.',
-    ollamaNotFound: 'Ollama executable not found on PATH.',
-    refreshModels: 'Refresh list',
-    noModels: 'No models installed. Pull some with “ollama pull”.',
-    roleRag: 'RAG generator', descRoleRag: 'Final answer in document mode',
-    roleChat: 'Sub-queries', descRoleChat: 'Conversation and query decomposition',
-    roleContextual: 'Contextual retrieval', descRoleContextual: 'Enriches chunks at indexing',
-    roleRecomp: 'RECOMP synthesis', descRoleRecomp: 'Summarizes context before generation',
-    modelSaveError: 'Could not change the model.',
-  },
-  ca: {
-    tabDocs: 'Documents', tabPipeline: 'Pipeline RAG',
-    corpusError: 'No s\'ha pogut canviar el magatzem.',
-    corpusConflict: 'Hi ha indexació en curs; espera un moment.',
-    corpusConnError: '✗ Error de connexió en canviar el corpus.',
-    collection: "Col·lecció ({n} docs)", noDocs: 'No hi ha documents indexats',
-    deleteDoc: 'Eliminar document', viewPdf: 'Veure PDF',
-    indexingFile: 'Indexant {file}', reindexingStatus: 'Re-indexació en curs...',
-    section1: '1. Indexació',
-    labelContextual: 'Recuperació contextual', descContextual: 'Enriqueix fragments amb LLM',
-    section2: '2. Recuperació',
-    labelHybrid: 'Cerca híbrida', descHybrid: 'Semàntica + BM25',
-    labelQueryDecomp: 'Descomposició de consultes', descQueryDecomp: 'Sub-consultes amb LLM auxiliar',
-    labelImageIndex: 'Indexat d\'imatges', descImageIndex: 'Recuperació visual directa',
-    labelImageDesc: 'Descripció de figures', descImageDesc: 'Text descriptiu per visió (cal reindexar i model amb visió)',
-    section3: '3. Rànquing i context',
-    labelReranker: 'Reordenador cross-encoder', descReranker: 'Reordenament de precisió',
-    labelExpandContext: 'Expandir context', descExpandContext: 'Afig fragments adjacents',
-    labelOptimizeContext: 'Optimitzar context', descOptimizeContext: 'Neteja artefactes PDF',
-    labelRecomp: 'Síntesi RECOMP', descRecomp: 'Sintetitza context amb LLM',
-    section4: '4. Re-indexació',
-    reindexHint: "Ajusta les opcions dalt i torna a indexar per a aplicar-les. Opcionalment afig PDFs nous.",
-    addPdfs: 'Afegir PDFs', remove: 'Llevar', reindexBtn: 'Re-indexar',
-    fragments: '{n} fragments', ollamaStatus: 'Ollama Local',
-    clearChat: 'Netejar xat', youLabel: 'Tu',
-    themeLight: 'Mode clar', themeDark: 'Mode fosc', close: 'Tancar',
-    sources: '{n} fonts', copyMsg: 'Copiar missatge',
-    placeholderRag: 'Pregunta sobre els teus documents…', placeholderChat: 'Escriu un missatge…',
-    footerMode: 'Mode document', footerModeChat: 'Mode conversa',
-    indexingTitle: "S'estan indexant els documents",
-    indexingHint: 'Pot tardar uns minuts depenent del teu maquinari.',
-    processing: 'Processant:', fileUnit: 'arxiu', fileUnitPlural: 'arxius',
-    autoRefresh: "La pàgina s'actualitzarà automàticament en acabar.",
-    retry: 'Reintentar', connErrorTitle: 'Error de connexió',
-    connecting: 'Connectant amb MonkeyGrab…',
-    reindexError: '✗ Error: {error}', reindexConnError: '✗ Error de connexió.',
-    confirmDelete: 'Eliminar "{name}"? S\'esborrarà de l\'índex i del disc.',
-    docDeleted: '✓ Document **{name}** eliminat.',
-    deleteError: '✗ Error en eliminar: {error}',
-    deleteConnError: '✗ Error de connexió en eliminar.',
-    unknownError: 'Error desconegut', queryError: 'Error en la consulta',
-    modelError: 'Error en la resposta del model',
-    connClosed: 'La connexió es va tancar abans de completar la resposta.',
-    settingsSaveError: "No s'ha pogut guardar l'ajust.",
-    noServer: "No s'ha pogut connectar amb el servidor. Està Flask executant-se?",
-    retryError: 'Error en reintentar',
-    indexingFailed: "La indexació no s'ha pogut completar.",
-    noResults: "No s'ha trobat informació rellevant als documents.",
-    tabModels: 'Models',
-    tabStudy: 'Estudiar',
-    modeStudy: 'ESTUDIAR',
-    studyKindSummary: 'Resum', studyKindOutline: 'Esquema', studyKindQuiz: 'Qüestionari',
-    studyPlaceholder: 'Escriu el nom d\'un document...',
-    studyHint: 'Tria què generar i escriu el document. Tab per autocompletar.',
-    studyNoMatch: 'No trobe cap document que coincidisca amb «{query}».',
-    studyAnswer: 'Resposta', studyShowAnswersInline: 'Veure respostes',
-    studyEmptyTitle: 'Estudia els teus documents',
-    studyEmptyBody: 'Un resum, un esquema o un qüestionari de qualsevol PDF indexat.',
-    studyPages: 'Pàgines', studyPagesShort: 'pàgines',
-    studyMalformed: 'El model no ha respectat el format. Prova una altra vegada, o amb un altre model.',
-    studyFailed: 'No s\'ha pogut generar: {error}',
-    storesLabel: 'Magatzem vectorial',
-    storeNotIndexed: 'sense indexar',
-    storeEmptyHint: 'Magatzem buit. Puja PDFs i reindexa per a activar-lo.',
-    fingerprintStaleWarning: "L'índex guardat no coincideix amb la configuració activa, però l'app continua funcionant amb ell. Reindexa quan vulgues: pot trigar una hora o més.",
-    modelsRoles: 'Rols de model',
-    ollamaTitle: 'Servidor Ollama',
-    ollamaOnline: 'En execució',
-    ollamaOffline: 'Aturat',
-    ollamaStartBtn: 'Arrancar Ollama',
-    ollamaStarting: 'Arrancant…',
-    ollamaStartFailed: "No s'ha pogut arrancar Ollama.",
-    ollamaNotFound: "No s'ha trobat l'executable d'Ollama al PATH.",
-    refreshModels: 'Actualitzar llista',
-    noModels: 'No hi ha models instal·lats. Descarrega\'n amb «ollama pull».',
-    roleRag: 'Generador RAG', descRoleRag: 'Resposta final en mode document',
-    roleChat: 'Subconsultes', descRoleChat: 'Conversa i descomposició de consultes',
-    roleContextual: 'Recuperació contextual', descRoleContextual: 'Enriqueix fragments en indexar',
-    roleRecomp: 'Síntesi RECOMP', descRoleRecomp: 'Resumeix el context abans de generar',
-    modelSaveError: 'No s\'ha pogut canviar el model.',
-  },
-} as const;
-
-const LANG_OPTIONS: Array<{ code: Lang; label: string }> = [
-  { code: 'es', label: 'ES' },
-  { code: 'en', label: 'EN' },
-  { code: 'ca', label: 'VAL' },
-];
-
-function normalizeLang(value: string | null): Lang {
-  return value === 'en' || value === 'ca' || value === 'es' ? value : 'es';
-}
-
-function fill(tpl: string, vars: Record<string, string | number>): string {
-  return tpl.replace(/\{(\w+)\}/g, (_, k) => String(vars[k] ?? ''));
-}
+import type {
+  Citation,
+  IndexingProgress,
+  Lang,
+  Message,
+  Mode,
+  ModelRole,
+  ModelRoles,
+  OllamaModel,
+  PipelineSettings,
+  StudyArtifact,
+  StudyKind,
+  VectorStore,
+} from './lib/types';
+import { fill, normalizeLang, STRINGS } from './lib/i18n';
+import { MarkdownContent } from './components/Markdown';
+import { StudyArtifactView } from './components/StudyArtifactView';
+import { ShimmerText } from './components/ShimmerText';
+import { PdfPane } from './components/PdfPane';
+import { LanguageToggle } from './components/LanguageToggle';
+import { ConnectionErrorScreen, IndexingScreen, LoadingScreen } from './components/StatusScreens';
+import { SettingsOverlay } from './components/SettingsOverlay';
 
 // =============================================================================
 // API Service — connects to Flask backend
@@ -541,169 +218,10 @@ async function streamSSE(
 }
 
 // =============================================================================
-// Small safe Markdown renderer
-// =============================================================================
-
-const MathInline: React.FC<{ tex: string }> = ({ tex }) => {
-  const html = katex.renderToString(tex, { throwOnError: false, displayMode: false });
-  return <span dangerouslySetInnerHTML={{ __html: html }} />;
-};
-
-const MathBlock: React.FC<{ tex: string }> = ({ tex }) => {
-  const html = katex.renderToString(tex, { throwOnError: false, displayMode: true });
-  return <div className="overflow-x-auto my-2" dangerouslySetInnerHTML={{ __html: html }} />;
-};
-
-function renderInlineMarkdown(text: string, keyPrefix: string): React.ReactNode[] {
-  const parts: React.ReactNode[] = [];
-  const pattern = /(\$\$[^$\n]+\$\$|\$(?!\$)(?:[^$\n\\]|\\.)+\$|`[^`]+`|\*\*[^*]+\*\*|\*[^*]+\*)/g;
-  let last = 0;
-  let match: RegExpExecArray | null;
-
-  while ((match = pattern.exec(text)) !== null) {
-    if (match.index > last) parts.push(text.slice(last, match.index));
-    const token = match[0];
-    const key = `${keyPrefix}-${match.index}`;
-
-    if (token.startsWith('$$')) {
-      parts.push(<MathBlock key={key} tex={token.slice(2, -2)} />);
-    } else if (token.startsWith('$')) {
-      parts.push(<MathInline key={key} tex={token.slice(1, -1)} />);
-    } else if (token.startsWith('`')) {
-      parts.push(<code key={key}>{token.slice(1, -1)}</code>);
-    } else if (token.startsWith('**')) {
-      parts.push(<strong key={key}>{token.slice(2, -2)}</strong>);
-    } else {
-      parts.push(<em key={key}>{token.slice(1, -1)}</em>);
-    }
-
-    last = match.index + token.length;
-  }
-
-  if (last < text.length) parts.push(text.slice(last));
-  return parts;
-}
-
-function MarkdownContent({ text, compact = false }: { text: string; compact?: boolean }) {
-  if (!text) return null;
-
-  const className = compact ? 'markdown-content compact' : 'markdown-content';
-
-  if (text.includes('```')) {
-    const nodes: React.ReactNode[] = [];
-    let inCode = false;
-    let codeLines: string[] = [];
-    let paragraph: string[] = [];
-    let inMath = false;
-    let mathLines: string[] = [];
-
-    const flushParagraph = (key: string) => {
-      if (!paragraph.length) return;
-      nodes.push(<p key={key}>{paragraph.map((line, i) => <React.Fragment key={i}>{i > 0 && <br />}{renderInlineMarkdown(line, `${key}-${i}`)}</React.Fragment>)}</p>);
-      paragraph = [];
-    };
-
-    text.split('\n').forEach((line, i) => {
-      if (line.trim() === '$$' && !inCode) {
-        if (inMath) {
-          flushParagraph(`p-${i}`);
-          nodes.push(<MathBlock key={`math-${i}`} tex={mathLines.join('\n')} />);
-          mathLines = [];
-          inMath = false;
-        } else {
-          flushParagraph(`p-${i}`);
-          inMath = true;
-        }
-        return;
-      }
-
-      if (inMath) {
-        mathLines.push(line);
-        return;
-      }
-
-      if (/^```/.test(line)) {
-        if (inCode) {
-          nodes.push(<pre key={`code-${i}`}><code>{codeLines.join('\n')}</code></pre>);
-          codeLines = [];
-          inCode = false;
-        } else {
-          flushParagraph(`p-${i}`);
-          inCode = true;
-        }
-        return;
-      }
-
-      if (inCode) {
-        codeLines.push(line);
-      } else if (line.trim()) {
-        paragraph.push(line);
-      } else {
-        flushParagraph(`p-${i}`);
-      }
-    });
-
-    if (inCode) nodes.push(<pre key="code-final"><code>{codeLines.join('\n')}</code></pre>);
-    flushParagraph('p-final');
-    return <div className={className}>{nodes}</div>;
-  }
-
-  return (
-    <div className={className}>
-      {text.split(/\n{2,}/).map((block, i) => {
-        const lines = block.split('\n').filter(Boolean);
-        if (!lines.length) return null;
-
-        const trimmed = block.trim();
-        if (trimmed.startsWith('$$') && trimmed.endsWith('$$') && trimmed.length > 4) {
-          const inner = trimmed.slice(2, -2).trim();
-          return <MathBlock key={i} tex={inner} />;
-        }
-
-        const heading = lines[0].match(/^(#{1,3})\s+(.+)$/);
-        if (heading) {
-          const Tag = heading[1].length === 1 ? 'h2' : heading[1].length === 2 ? 'h3' : 'h4';
-          return <Tag key={i}>{renderInlineMarkdown(heading[2], `h-${i}`)}</Tag>;
-        }
-
-        if (lines.every(line => /^[-*]\s+/.test(line))) {
-          return (
-            <ul key={i}>
-              {lines.map((line, j) => (
-                <li key={j}>{renderInlineMarkdown(line.replace(/^[-*]\s+/, ''), `ul-${i}-${j}`)}</li>
-              ))}
-            </ul>
-          );
-        }
-
-        if (lines.every(line => /^\d+\.\s+/.test(line))) {
-          return (
-            <ol key={i}>
-              {lines.map((line, j) => (
-                <li key={j}>{renderInlineMarkdown(line.replace(/^\d+\.\s+/, ''), `ol-${i}-${j}`)}</li>
-              ))}
-            </ol>
-          );
-        }
-
-        return (
-          <p key={i}>
-            {lines.map((line, j) => (
-              <React.Fragment key={j}>
-                {j > 0 && <br />}
-                {renderInlineMarkdown(line, `p-${i}-${j}`)}
-              </React.Fragment>
-            ))}
-          </p>
-        );
-      })}
-    </div>
-  );
-}
-
-// =============================================================================
 // App Component
 // =============================================================================
+
+const GREETING: Record<Lang, string> = { es: 'Hola', en: 'Hi', ca: 'Hola' };
 
 export default function App() {
   const [lang, setLangState] = useState<Lang>(() =>
@@ -1330,253 +848,40 @@ export default function App() {
   if (isIndexing) {
     const showRetry = indexingError && /falló|fallido|failed|refused|no se pudo/i.test(indexingError);
     return (
-      <div className="flex h-screen items-center justify-center bg-transparent text-zinc-300 p-4">
-        <div className="glass-panel rounded-3xl p-10 max-w-md text-center space-y-4">
-          <Loader2 className="w-12 h-12 text-orange-500 animate-spin mx-auto" />
-          <h2 className="text-xl font-semibold text-white">
-            {T.indexingTitle}
-          </h2>
-          <p className="text-zinc-400 text-sm">
-            {T.indexingHint}
-          </p>
-          {indexingError && showRetry && (
-            <p className="text-amber-400 text-sm">{indexingError}</p>
-          )}
-          {indexingProgress ? (
-            <div className="space-y-1">
-              <p className="text-zinc-300 text-sm font-medium">
-                {T.processing} <span className="text-orange-500">{indexingProgress.file}</span>
-              </p>
-              <p className="text-zinc-500 text-xs">
-                {indexingProgress.file_index} / {indexingProgress.total_files} {indexingProgress.total_files !== 1 ? T.fileUnitPlural : T.fileUnit}
-              </p>
-              <div className="w-full bg-zinc-800 rounded-full h-1.5 mt-2">
-                <div
-                  className="bg-orange-500 h-1.5 rounded-full transition-all duration-500"
-                  style={{ width: `${(indexingProgress.file_index / indexingProgress.total_files) * 100}%` }}
-                />
-              </div>
-            </div>
-          ) : (
-            <p className="text-zinc-500 text-xs">
-              {T.autoRefresh}
-            </p>
-          )}
-          {showRetry && (
-            <button
-              className="px-6 py-2 bg-[var(--accent)] text-[var(--accent-contrast)] font-semibold hover:bg-[var(--accent-hover)] transition-colors"
-              onClick={handleRetry}
-            >
-              {T.retry}
-            </button>
-          )}
-        </div>
-      </div>
+      <IndexingScreen
+        title={T.indexingTitle}
+        hint={T.indexingHint}
+        error={indexingError}
+        showRetry={Boolean(showRetry)}
+        progress={indexingProgress}
+        processingLabel={T.processing}
+        fileUnit={T.fileUnit}
+        fileUnitPlural={T.fileUnitPlural}
+        autoRefreshLabel={T.autoRefresh}
+        retryLabel={T.retry}
+        onRetry={handleRetry}
+      />
     );
   }
 
   // ---- Connection error screen ----
   if (initError) {
     return (
-      <div className="flex h-screen items-center justify-center bg-transparent text-zinc-300 p-4">
-        <div className="glass-panel rounded-3xl p-10 max-w-md text-center space-y-4">
-          <AlertCircle className="w-12 h-12 text-red-400 mx-auto" />
-          <h2 className="text-xl font-semibold text-white">{T.connErrorTitle}</h2>
-          <p className="text-zinc-400 text-sm">{initError}</p>
-          <button
-            className="px-6 py-2 bg-[var(--accent)] text-[var(--accent-contrast)] font-semibold hover:bg-[var(--accent-hover)] transition-colors"
-            onClick={handleRetry}
-          >
-            {T.retry}
-          </button>
-        </div>
-      </div>
+      <ConnectionErrorScreen
+        title={T.connErrorTitle}
+        error={initError}
+        retryLabel={T.retry}
+        onRetry={handleRetry}
+      />
     );
   }
 
   // ---- Loading screen ----
   if (!isInitialized) {
     return (
-      <div className="flex h-screen items-center justify-center bg-transparent text-zinc-300">
-        <div className="flex flex-col items-center gap-4">
-          <Loader2 className="w-8 h-8 text-orange-400 animate-spin" />
-          <p className="text-zinc-500 text-sm">{T.connecting}</p>
-        </div>
-      </div>
+      <LoadingScreen message={T.connecting} />
     );
   }
-
-  // ---- Overlay panels (rendered full-area in the main column) ----
-  const renderModelsPanel = () => (
-    <div className="mx-auto w-full max-w-4xl space-y-4">
-      {/* Ollama server status */}
-      <div className={`border p-3 ${ollamaStatus.running ? 'border-[var(--border)] bg-[var(--surface)]' : 'border-amber-500/30 bg-amber-500/10'}`}>
-        <div className="flex items-center justify-between gap-2">
-          <div className="flex min-w-0 items-center gap-2">
-            <span className={`h-1.5 w-1.5 flex-shrink-0 rounded-full ${ollamaStatus.running ? 'bg-green-500' : 'bg-red-500'}`} />
-            <div className="min-w-0">
-              <div className="t-h3 text-[var(--text)]">{T.ollamaTitle}</div>
-              <div className="truncate t-body-sm text-[var(--text-muted)]">
-                {ollamaStatus.running ? T.ollamaOnline : T.ollamaOffline}
-              </div>
-            </div>
-          </div>
-          {ollamaStatus.running ? (
-            <button
-              type="button"
-              className="flex-shrink-0 p-2 text-[var(--text-muted)] transition-all hover:bg-[var(--surface-2)] hover:text-[var(--accent)]"
-              onClick={refreshOllama}
-              title={T.refreshModels}
-            >
-              <RefreshCw className="h-4 w-4" />
-            </button>
-          ) : (
-            <button
-              type="button"
-              className="flex flex-shrink-0 items-center gap-1.5 border border-amber-500/40 bg-amber-500/15 px-3 py-1.5 text-[11px] font-semibold text-amber-500 transition-all hover:bg-amber-500/25 disabled:opacity-50"
-              onClick={handleStartOllama}
-              disabled={ollamaStarting}
-            >
-              {ollamaStarting ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <Power className="h-3.5 w-3.5" />}
-              {ollamaStarting ? T.ollamaStarting : T.ollamaStartBtn}
-            </button>
-          )}
-        </div>
-      </div>
-
-      {modelError && (
-        <div className="border border-red-500/20 bg-red-500/10 px-3 py-2 text-xs text-red-400">{modelError}</div>
-      )}
-
-      {/* Model role selectors — two columns in the wide overlay */}
-      <div className="space-y-2">
-        <div className="t-label text-[var(--text-muted)] pl-1">{T.modelsRoles}</div>
-        {ollamaStatus.running && ollamaModels.length === 0 ? (
-          <p className="px-1 py-3 text-xs text-[var(--text-muted)]">{T.noModels}</p>
-        ) : (
-          <div className="grid gap-2 sm:grid-cols-2">
-            {([
-              { role: 'rag' as ModelRole, label: T.roleRag, desc: T.descRoleRag },
-              { role: 'chat' as ModelRole, label: T.roleChat, desc: T.descRoleChat },
-              { role: 'contextual' as ModelRole, label: T.roleContextual, desc: T.descRoleContextual },
-              { role: 'recomp' as ModelRole, label: T.roleRecomp, desc: T.descRoleRecomp },
-            ]).map(({ role, label, desc }) => {
-              const current = modelRoles?.[role] ?? '';
-              const names = ollamaModels.map(m => m.name);
-              const options = current && !names.includes(current) ? [current, ...names] : names;
-              return (
-                <div key={role} className="border border-[var(--border)] bg-[var(--surface)] p-2.5">
-                  <div className="mb-1 flex items-center justify-between gap-2">
-                    <span className="t-h3 text-[var(--accent)]">{label}</span>
-                    {savingRole === role && <Loader2 className="h-3 w-3 animate-spin text-[var(--accent)]" />}
-                  </div>
-                  <p className="mb-2 t-body-sm text-[var(--text-muted)]">{desc}</p>
-                  <ModelSelect
-                    value={current}
-                    options={options}
-                    disabled={!ollamaStatus.running || savingRole !== null}
-                    onChange={v => handleRoleChange(role, v)}
-                  />
-                </div>
-              );
-            })}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-
-  const renderPipelinePanel = () => (
-    <div className="mx-auto w-full max-w-4xl space-y-4">
-      {(settingsError || isReindexing) && (
-        <div className={`border px-3 py-2 text-xs ${settingsError ? 'border-red-500/20 bg-red-500/10 text-red-400' : 'border-[var(--accent)]/30 bg-[var(--accent)]/10 text-[var(--accent)]'}`}>
-          {settingsError ? (
-            <span>{settingsError}</span>
-          ) : indexingProgress ? (
-            <div className="space-y-1">
-              <div className="flex items-center justify-between gap-3">
-                <span className="truncate">{fill(T.indexingFile, { file: indexingProgress.file })}</span>
-                <span className="font-mono text-[10px] text-[var(--accent)]">{indexingProgress.file_index}/{indexingProgress.total_files}</span>
-              </div>
-              <div className="h-1.5 bg-[var(--surface-2)] overflow-hidden">
-                <div
-                  className="h-full bg-[var(--accent)] transition-all duration-500"
-                  style={{ width: `${Math.max(5, (indexingProgress.file_index / indexingProgress.total_files) * 100)}%` }}
-                />
-              </div>
-            </div>
-          ) : (
-            <span>{T.reindexingStatus}</span>
-          )}
-        </div>
-      )}
-
-      <div className="grid gap-4 md:grid-cols-2">
-        <section className="flex flex-col border border-[var(--border)] bg-[var(--popover)] p-3">
-          <h3 className="t-label mb-3 text-[var(--accent)]">{T.section1}</h3>
-          <div className="space-y-0.5">
-            <Toggle label={T.labelContextual} checked={settings.contextualRetrieval} onChange={() => toggleSetting('contextualRetrieval')} desc={T.descContextual} />
-            <Toggle label={T.labelImageIndex} checked={settings.imageIndexing} onChange={() => toggleSetting('imageIndexing')} desc={T.descImageIndex} />
-            <Toggle label={T.labelImageDesc} checked={settings.imageDescription} onChange={() => toggleSetting('imageDescription')} desc={T.descImageDesc} />
-          </div>
-        </section>
-
-        <section className="flex flex-col border border-[var(--border)] bg-[var(--popover)] p-3">
-          <h3 className="t-label mb-3 text-[var(--accent)]">{T.section2}</h3>
-          <div className="space-y-0.5">
-            <Toggle label={T.labelHybrid} checked={settings.hybridSearch} onChange={() => toggleSetting('hybridSearch')} desc={T.descHybrid} />
-            <Toggle label={T.labelQueryDecomp} checked={settings.queryDecomposition} onChange={() => toggleSetting('queryDecomposition')} desc={T.descQueryDecomp} />
-          </div>
-        </section>
-
-        <section className="flex flex-col border border-[var(--border)] bg-[var(--popover)] p-3 md:col-span-2">
-          <h3 className="t-label mb-3 text-[var(--accent)]">{T.section3}</h3>
-          <div className="grid gap-0 sm:grid-cols-2">
-            <Toggle label={T.labelReranker} checked={settings.reranker} onChange={() => toggleSetting('reranker')} desc={T.descReranker} />
-            <Toggle label={T.labelExpandContext} checked={settings.expandContext} onChange={() => toggleSetting('expandContext')} desc={T.descExpandContext} />
-            <Toggle label={T.labelOptimizeContext} checked={settings.optimizeContext} onChange={() => toggleSetting('optimizeContext')} desc={T.descOptimizeContext} />
-            <Toggle label={T.labelRecomp} checked={settings.recompSynthesis} onChange={() => toggleSetting('recompSynthesis')} desc={T.descRecomp} />
-          </div>
-        </section>
-
-        <section className="flex flex-col border border-[var(--border)] bg-[var(--popover)] p-4 md:col-span-2">
-          <h3 className="t-label mb-2 text-[var(--accent)]">{T.section4}</h3>
-          <p className="mb-4 text-xs text-[var(--text-muted)]">{T.reindexHint}</p>
-          <div className="flex flex-col gap-3 sm:flex-row sm:items-stretch">
-            <div className="flex flex-1 gap-2">
-              <button
-                className="flex flex-1 flex-col items-center justify-center gap-1 border border-dashed border-[var(--border-strong)] px-4 py-3 text-sm text-[var(--text-muted)] transition-all hover:border-[var(--accent)] hover:bg-[var(--popover-hover)] hover:text-[var(--text)] disabled:opacity-50 group"
-                onClick={() => reindexFileInputRef.current?.click()}
-                disabled={isReindexing}
-              >
-                <FileUp className="h-5 w-5 group-hover:text-[var(--accent)]" />
-                <span className="font-medium">
-                  {pendingReindexFiles.length ? `${pendingReindexFiles.length} PDF(s)` : T.addPdfs}
-                </span>
-              </button>
-              {pendingReindexFiles.length > 0 && (
-                <button
-                  className="px-3 text-[var(--text-muted)] transition-all hover:bg-red-500/10 hover:text-red-400"
-                  onClick={() => { setPendingReindexFiles([]); if (reindexFileInputRef.current) reindexFileInputRef.current.value = ''; }}
-                  title={T.remove}
-                >
-                  <X className="h-5 w-5" />
-                </button>
-              )}
-            </div>
-            <button
-              className="flex items-center justify-center gap-2 bg-[var(--accent)] px-6 py-3 text-sm font-semibold text-[var(--accent-contrast)] transition-all hover:bg-[var(--accent-hover)] disabled:opacity-50 sm:min-w-[10rem]"
-              onClick={handleReindex}
-              disabled={isReindexing}
-            >
-              {isReindexing ? <Loader2 className="h-5 w-5 animate-spin" /> : <RefreshCw className="h-5 w-5" />}
-              {T.reindexBtn}
-            </button>
-          </div>
-        </section>
-      </div>
-    </div>
-  );
 
   // ---- Main UI ----
   return (
@@ -1763,28 +1068,29 @@ export default function App() {
 
         {/* Models / Pipeline overlay — occupies the full main area, like the PDF viewer */}
         {mainPanel && pdfViewer?.mode !== 'full' && (
-          <div className="flex-1 flex flex-col min-w-0 relative">
-            <header className="h-20 border-b border-[var(--border)] flex items-center justify-between gap-3 px-6 bg-[var(--surface)] z-10">
-              <div className="flex items-center gap-2 min-w-0">
-                {mainPanel === 'models'
-                  ? <Ollama className="w-5 h-5 flex-shrink-0 text-[var(--text)]" />
-                  : <Database className="w-5 h-5 flex-shrink-0" />}
-                <h2 className="t-h2 text-[var(--text)] truncate">{mainPanel === 'models' ? T.tabModels : T.tabPipeline}</h2>
-              </div>
-              <button
-                type="button"
-                onClick={() => setMainPanel(null)}
-                className="p-2 text-[var(--text-muted)] hover:text-[var(--text)] bg-[var(--surface)] hover:bg-[var(--surface-2)] border border-[var(--border)] transition-colors flex-shrink-0"
-                title={T.close}
-                aria-label={T.close}
-              >
-                <X className="w-4 h-4" />
-              </button>
-            </header>
-            <div className="flex-1 overflow-y-auto p-6 md:p-8 custom-scrollbar">
-              {mainPanel === 'models' ? renderModelsPanel() : renderPipelinePanel()}
-            </div>
-          </div>
+          <SettingsOverlay
+            mainPanel={mainPanel}
+            onClose={() => setMainPanel(null)}
+            ollamaStatus={ollamaStatus}
+            ollamaStarting={ollamaStarting}
+            refreshOllama={refreshOllama}
+            handleStartOllama={handleStartOllama}
+            ollamaModels={ollamaModels}
+            modelRoles={modelRoles}
+            savingRole={savingRole}
+            modelError={modelError}
+            handleRoleChange={handleRoleChange}
+            settings={settings}
+            toggleSetting={toggleSetting}
+            settingsError={settingsError}
+            isReindexing={isReindexing}
+            indexingProgress={indexingProgress}
+            pendingReindexFiles={pendingReindexFiles}
+            setPendingReindexFiles={setPendingReindexFiles}
+            reindexFileInputRef={reindexFileInputRef}
+            handleReindex={handleReindex}
+            strings={T}
+          />
         )}
 
         {/* Chat column — hidden while a full-screen PDF or a panel is open */}
@@ -1918,7 +1224,7 @@ export default function App() {
                             kind={msg.artifactKind}
                             revealed={!!revealedQuiz[msg.id]}
                             onReveal={() => setRevealedQuiz(prev => ({ ...prev, [msg.id]: true }))}
-                            strings={T as unknown as Record<string, string>}
+                            strings={T}
                           />
                         ) : msg.content ? (
                           <MarkdownContent text={msg.content} />
@@ -2082,331 +1388,6 @@ export default function App() {
         </div>
         )}
       </main>
-    </div>
-  );
-}
-
-// =============================================================================
-// Study artifact — rendered in the chat bubble, one shape per kind
-// =============================================================================
-
-function OutlineTree({ nodes, depth = 0 }: { nodes: OutlineNode[]; depth?: number }) {
-  // Top level is a stack of headings; every level below hangs off a rule.
-  // Indentation alone did not read as hierarchy — 16px against 15px text is
-  // noise, and the result looked like a flat list. The rule is what makes a
-  // child visibly belong to its parent.
-  return (
-    <ul className={depth === 0 ? 'space-y-3' : 'mt-1.5 space-y-1.5 border-l border-[var(--border)] pl-4'}>
-      {nodes.map((node, i) => (
-        <li key={`${depth}-${i}-${node.title}`} className="relative">
-          {depth > 0 && (
-            // Ticks the rule at each child, so siblings are countable.
-            <span className="absolute -left-4 top-[0.7em] w-2 h-px bg-[var(--border)]" aria-hidden="true" />
-          )}
-          <span
-            className={
-              depth === 0
-                ? 'block text-[var(--text)] font-semibold text-[15px]'
-                : depth === 1
-                  ? 'block text-[var(--text)] text-[14px]'
-                  : 'block text-[var(--text-muted)] text-[13px]'
-            }
-          >
-            {node.title}
-          </span>
-          {node.children?.length ? <OutlineTree nodes={node.children} depth={depth + 1} /> : null}
-        </li>
-      ))}
-    </ul>
-  );
-}
-
-function StudyArtifactView({
-  artifact, kind, revealed, onReveal, strings,
-}: {
-  artifact: StudyArtifact;
-  kind: StudyKind;
-  revealed: boolean;
-  onReveal: () => void;
-  strings: Record<string, string>;
-}) {
-  const pages = Array.from(new Set<number>([
-    ...(artifact.sections ?? []).flatMap(x => x.source_pages ?? []),
-    ...(artifact.questions ?? []).flatMap(x => x.source_pages ?? []),
-  ])).sort((a, b) => a - b);
-
-  return (
-    <div className="w-full space-y-4 text-[15px] leading-relaxed">
-      {kind === 'summary' && (artifact.sections ?? []).map((section, i) => (
-        <div key={i} className="space-y-1">
-          {section.heading && (
-            <h3 className="text-[var(--text)] font-semibold">{section.heading}</h3>
-          )}
-          <div className="text-[var(--text-muted)]">
-            <MarkdownContent text={section.body} />
-          </div>
-        </div>
-      ))}
-
-      {kind === 'outline' && <OutlineTree nodes={artifact.nodes ?? []} />}
-
-      {kind === 'quiz' && (
-        <div className="space-y-5">
-          {(artifact.questions ?? []).map((q, i) => (
-            <div key={i} className="space-y-2">
-              <p className="text-[var(--text)] font-semibold">
-                <span className="text-[var(--accent)] mr-1.5">{i + 1}.</span>{q.prompt}
-              </p>
-              <ul className="space-y-1.5">
-                {q.options.map((option, j) => {
-                  const isKey = revealed && j === q.correct_index;
-                  return (
-                    <li
-                      key={j}
-                      className={`flex items-start gap-2.5 px-3 py-1.5 border transition-colors ${
-                        isKey
-                          ? 'border-[var(--accent)]/40 bg-[var(--accent)]/10 text-[var(--text)]'
-                          : 'border-transparent text-[var(--text-muted)]'
-                      }`}
-                    >
-                      <span className={`font-mono text-xs mt-0.5 ${isKey ? 'text-[var(--accent)]' : 'text-[var(--text-faint)]'}`}>
-                        {String.fromCharCode(97 + j)}
-                      </span>
-                      <span className="flex-1">{option}</span>
-                      {isKey && (
-                        <CheckCircle2 className="w-4 h-4 shrink-0 mt-0.5 text-[var(--accent)]" />
-                      )}
-                    </li>
-                  );
-                })}
-              </ul>
-            </div>
-          ))}
-          {/* Hidden by default: a quiz whose key is on screen from the first
-              render is a summary with extra steps. */}
-          {!revealed && (artifact.questions ?? []).length > 0 && (
-            <button
-              type="button"
-              onClick={onReveal}
-              className="px-3 py-1.5 text-[11px] font-bold tracking-wide uppercase text-[var(--text-muted)] hover:text-[var(--text)] bg-[var(--surface)] hover:bg-[var(--surface-2)] border border-[var(--border)] transition-colors"
-            >
-              {strings.studyShowAnswersInline}
-            </button>
-          )}
-        </div>
-      )}
-
-      {artifact.source_document && (
-        <div className="flex flex-wrap items-center gap-2 pt-1 text-xs text-[var(--text-faint)]">
-          <span className="inline-flex items-center gap-1.5">
-            <FileText className="w-3.5 h-3.5" />
-            {artifact.source_document}
-          </span>
-          {pages.length > 0 && <span>· {pages.length} {strings.studyPagesShort}</span>}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// =============================================================================
-// PDF Viewer Modal
-// =============================================================================
-
-const GREETING: Record<Lang, string> = { es: 'Hola', en: 'Hi', ca: 'Hola' };
-
-/** Letters cycle white → accent → white (staggered): a colour shimmer, no motion.
-    Render inside a flex container that sets the font size and weight. */
-function ShimmerText({ text }: { text: string }) {
-  return (
-    <>
-      {Array.from(text).map((ch, i) => (
-        <span
-          key={i}
-          className="shimmer-char"
-          style={{ animationDelay: `${i * 0.12}s` }}
-        >
-          {ch}
-        </span>
-      ))}
-    </>
-  );
-}
-
-function PdfPane({ doc, page, onClose }: { doc: string; page: number; onClose: () => void }) {
-  const base = `/api/pdf/${encodeURIComponent(doc)}`;
-  // toolbar=0 hides the browser's built-in PDF bar (which repeats the file name).
-  const src = `${base}#page=${page}&toolbar=0`;
-  const [loaded, setLoaded] = useState(false);
-
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => { if (e.key === 'Escape') onClose(); };
-    window.addEventListener('keydown', onKey);
-    return () => window.removeEventListener('keydown', onKey);
-  }, [onClose]);
-
-  // Reset the loading overlay whenever the target document/page changes.
-  useEffect(() => { setLoaded(false); }, [src]);
-
-  return (
-    <div className="flex flex-col w-full h-full bg-[#0d0c12]">
-      <div className="h-20 flex items-center justify-between px-5 bg-white/[0.03] border-b border-white/5 flex-shrink-0">
-        <div className="flex items-center gap-3 min-w-0">
-          <FileText className="w-4 h-4 text-orange-400 flex-shrink-0" />
-          <span className="text-sm font-medium text-zinc-200 truncate">{doc}</span>
-        </div>
-        <div className="flex items-center gap-1 flex-shrink-0">
-          {/* Fallback: open in the system browser if the embedded viewer stays blank. */}
-          <a
-            href={base}
-            target="_blank"
-            rel="noreferrer"
-            className="p-2 rounded-full text-zinc-500 hover:text-orange-400 hover:bg-white/10 transition-all"
-            title="Abrir en el navegador"
-          >
-            <Eye className="w-5 h-5" />
-          </a>
-          <button
-            onClick={onClose}
-            className="p-2 rounded-full text-zinc-500 hover:text-white hover:bg-white/10 transition-all"
-            title="Cerrar (Esc)"
-          >
-            <X className="w-5 h-5" />
-          </button>
-        </div>
-      </div>
-      <div className="relative flex-1 min-h-0">
-        {!loaded && (
-          <div className="absolute inset-0 z-10 flex items-center justify-center bg-[#0d0c12]">
-            <Loader2 className="w-8 h-8 text-orange-500 animate-spin" />
-          </div>
-        )}
-        <iframe
-          key={src}
-          src={src}
-          onLoad={() => setLoaded(true)}
-          className="w-full h-full border-none bg-zinc-900"
-          title={doc}
-        />
-      </div>
-    </div>
-  );
-}
-
-// =============================================================================
-// Toggle Component
-// =============================================================================
-
-function Toggle({ label, checked, onChange, desc }: { label: string; checked: boolean; onChange: () => void; desc: string }) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={checked}
-      className="flex w-full items-center justify-between gap-4 rounded-lg p-2 text-left transition-colors hover:bg-[var(--popover-hover)] focus:outline-none focus:ring-2 focus:ring-orange-500/50 group"
-      onClick={onChange}
-    >
-      <span className="flex-1">
-        <span className="block text-sm font-medium text-[var(--text)] transition-colors">{label}</span>
-        <span className="mt-1 block text-[11px] leading-snug text-[var(--text-muted)]">{desc}</span>
-      </span>
-      <span className={`relative inline-flex h-6 w-11 flex-shrink-0 items-center justify-center rounded-full transition-colors duration-300 ease-in-out ${checked ? 'bg-orange-500 shadow-[0_0_10px_rgba(230,140,82,0.4)]' : 'bg-[var(--popover-hover)]'}`}>
-        <span className={`pointer-events-none inline-block h-4 w-4 transform rounded-full bg-white shadow-md transition duration-300 ease-in-out ${checked ? 'translate-x-2.5' : '-translate-x-2.5'}`} />
-      </span>
-    </button>
-  );
-}
-
-function LanguageToggle({ lang, setLang }: { lang: Lang; setLang: (lang: Lang) => void }) {
-  return (
-    <div className="flex items-center gap-1 border border-[var(--border)] bg-[var(--surface)] p-1">
-      <Languages className="ml-1.5 h-3.5 w-3.5 text-[var(--text-faint)]" />
-      {LANG_OPTIONS.map(option => (
-        <button
-          key={option.code}
-          type="button"
-          className={`w-9 text-center px-1 py-1 text-[10px] font-bold tracking-wide transition-all ${
-            lang === option.code
-              ? 'bg-[var(--accent)] text-[var(--accent-contrast)]'
-              : 'text-[var(--text-muted)] hover:bg-[var(--surface-2)] hover:text-[var(--text)]'
-          }`}
-          onClick={() => setLang(option.code)}
-          aria-pressed={lang === option.code}
-        >
-          {option.label}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-// Styled replacement for the native <select> used to pick a model per role.
-// Uses semantic theme tokens so the trigger and panel match light/dark mode.
-function ModelSelect({
-  value, options, disabled, onChange,
-}: {
-  value: string;
-  options: string[];
-  disabled?: boolean;
-  onChange: (value: string) => void;
-}) {
-  const [open, setOpen] = useState(false);
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!open) return;
-    const onDown = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node)) setOpen(false);
-    };
-    const onKey = (e: KeyboardEvent) => { if (e.key === 'Escape') setOpen(false); };
-    document.addEventListener('mousedown', onDown);
-    document.addEventListener('keydown', onKey);
-    return () => {
-      document.removeEventListener('mousedown', onDown);
-      document.removeEventListener('keydown', onKey);
-    };
-  }, [open]);
-
-  return (
-    <div ref={ref} className="relative">
-      <button
-        type="button"
-        disabled={disabled}
-        onClick={() => setOpen(o => !o)}
-        className="flex w-full items-center justify-between gap-2 rounded-xl border border-[var(--border)] bg-[var(--popover)] px-2.5 py-2 text-xs text-[var(--text)] transition-colors hover:border-[var(--border-strong)] focus:border-orange-500/50 focus:outline-none focus:ring-2 focus:ring-orange-500/20 disabled:opacity-50"
-      >
-        <span className="truncate">{value || '—'}</span>
-        <ChevronDown className={`h-3.5 w-3.5 flex-shrink-0 text-[var(--text-faint)] transition-transform ${open ? 'rotate-180' : ''}`} />
-      </button>
-      <AnimatePresence>
-        {open && (
-          <motion.ul
-            initial={{ opacity: 0, y: -4 }}
-            animate={{ opacity: 1, y: 0 }}
-            exit={{ opacity: 0, y: -4 }}
-            transition={{ duration: 0.15 }}
-            className="custom-scrollbar absolute z-50 mt-1.5 max-h-56 w-full overflow-y-auto rounded-xl border border-[var(--border)] bg-[var(--popover)] p-1 shadow-xl"
-          >
-            {options.map(name => {
-              const selected = name === value;
-              return (
-                <li key={name}>
-                  <button
-                    type="button"
-                    onClick={() => { onChange(name); setOpen(false); }}
-                    className={`flex w-full items-center justify-between gap-2 rounded-lg px-2.5 py-1.5 text-left text-xs transition-colors ${
-                      selected ? 'bg-[var(--popover-active)] text-[var(--accent)]' : 'text-[var(--text)] hover:bg-[var(--popover-hover)]'
-                    }`}
-                  >
-                    <span className="truncate">{name}</span>
-                    {selected && <Check className="h-3.5 w-3.5 flex-shrink-0 text-[var(--accent)]" />}
-                  </button>
-                </li>
-              );
-            })}
-          </motion.ul>
-        )}
-      </AnimatePresence>
     </div>
   );
 }
