+++ rag/web/frontend/src/lib/i18n.ts
@@ -1,0 +1,266 @@
+import type { Lang } from './types';
+
+export const STRINGS = {
+  es: {
+    tabDocs: 'Documentos', tabPipeline: 'Pipeline RAG',
+    corpusError: 'No se pudo cambiar el almacén.',
+    corpusConflict: 'Hay indexación en curso; espera un momento.',
+    corpusConnError: '✗ Error de conexión al cambiar el corpus.',
+    collection: 'Colección ({n} docs)', noDocs: 'No hay documentos indexados',
+    deleteDoc: 'Eliminar documento', viewPdf: 'Ver PDF',
+    indexingFile: 'Indexando {file}', reindexingStatus: 'Re-indexación en curso...',
+    section1: '1. Indexación',
+    labelContextual: 'Recuperación contextual', descContextual: 'Enriquece fragmentos con LLM',
+    section2: '2. Recuperación',
+    labelHybrid: 'Búsqueda híbrida', descHybrid: 'Semántica + BM25',
+    labelQueryDecomp: 'Descomposición de consultas', descQueryDecomp: 'Subconsultas con LLM auxiliar',
+    labelImageIndex: 'Indexado de imágenes', descImageIndex: 'Recuperación visual directa',
+    labelImageDesc: 'Descripción de figuras', descImageDesc: 'Texto descriptivo por visión (requiere reindexado y modelo con visión)',
+    section3: '3. Ranking y contexto',
+    labelReranker: 'Reordenador cross-encoder', descReranker: 'Reordenamiento de precisión',
+    labelExpandContext: 'Expandir contexto', descExpandContext: 'Añade fragmentos adyacentes',
+    labelOptimizeContext: 'Optimizar contexto', descOptimizeContext: 'Limpia artefactos PDF',
+    labelRecomp: 'Síntesis RECOMP', descRecomp: 'Sintetiza contexto con LLM',
+    section4: '4. Reindexación',
+    reindexHint: 'Ajusta las opciones arriba y reindexa para aplicarlas. Opcionalmente añade PDFs nuevos.',
+    addPdfs: 'Añadir PDFs', remove: 'Quitar', reindexBtn: 'Reindexar',
+    fragments: '{n} fragmentos', ollamaStatus: 'Ollama Local',
+    clearChat: 'Limpiar chat', youLabel: 'Tú',
+    themeLight: 'Modo claro', themeDark: 'Modo oscuro', close: 'Cerrar',
+    sources: '{n} fuentes', copyMsg: 'Copiar mensaje',
+    placeholderRag: 'Pregunta sobre tus documentos…', placeholderChat: 'Escribe un mensaje…',
+    footerMode: 'Modo documento', footerModeChat: 'Modo conversación',
+    indexingTitle: 'Se están indexando los documentos',
+    indexingHint: 'Puede tardar unos minutos dependiendo de tu hardware.',
+    processing: 'Procesando:', fileUnit: 'archivo', fileUnitPlural: 'archivos',
+    autoRefresh: 'La página se actualizará automáticamente al terminar.',
+    retry: 'Reintentar', connErrorTitle: 'Error de conexión',
+    connecting: 'Conectando con MonkeyGrab…',
+    reindexError: '✗ Error: {error}', reindexConnError: '✗ Error de conexión.',
+    confirmDelete: '¿Eliminar "{name}"? Se borrará del índice y del disco.',
+    docDeleted: '✓ Documento **{name}** eliminado.',
+    deleteError: '✗ Error al eliminar: {error}',
+    deleteConnError: '✗ Error de conexión al eliminar.',
+    unknownError: 'Error desconocido', queryError: 'Error en la consulta',
+    modelError: 'Error en la respuesta del modelo',
+    connClosed: 'La conexión se cerró antes de completar la respuesta.',
+    settingsSaveError: 'No se pudo guardar el ajuste.',
+    noServer: 'No se pudo conectar con el servidor. ¿Está Flask ejecutándose?',
+    retryError: 'Error al reintentar',
+    indexingFailed: 'La indexación no pudo completarse.',
+    noResults: 'No se encontró información relevante en los documentos.',
+    tabModels: 'Modelos',
+    modeStudy: 'ESTUDIAR',
+    studyKindSummary: 'Resumen', studyKindOutline: 'Esquema', studyKindQuiz: 'Cuestionario',
+    studyPlaceholder: 'Escribe el nombre de un documento...',
+    studyHint: 'Elige qué generar y escribe el documento. Tab para autocompletar.',
+    studyNoMatch: 'No encuentro ningún documento que coincida con «{query}».',
+    studyAnswer: 'Respuesta', studyShowAnswersInline: 'Ver respuestas',
+    studyEmptyTitle: 'Estudia tus documentos',
+    studyEmptyBody: 'Un resumen, un esquema o un cuestionario de cualquier PDF indexado.',
+    studyPages: 'Páginas', studyPagesShort: 'páginas',
+    studyMalformed: 'El modelo no respetó el formato. Prueba otra vez, o con otro modelo.',
+    studyFailed: 'No se pudo generar: {error}',
+    storesLabel: 'Almacén vectorial',
+    storeNotIndexed: 'sin indexar',
+    storeEmptyHint: 'Almacén vacío. Sube PDFs y reindexa para activarlo.',
+    fingerprintStaleWarning: 'El índice guardado no coincide con la configuración activa, pero la app sigue funcionando con él. Reindexa cuando quieras: puede tardar una hora o más.',
+    modelsRoles: 'Roles de modelo',
+    ollamaTitle: 'Servidor Ollama',
+    ollamaOnline: 'En ejecución',
+    ollamaOffline: 'Detenido',
+    ollamaStartBtn: 'Arrancar Ollama',
+    ollamaStarting: 'Arrancando…',
+    ollamaStartFailed: 'No se pudo arrancar Ollama.',
+    ollamaNotFound: 'No se encontró el ejecutable de Ollama en el PATH.',
+    refreshModels: 'Actualizar lista',
+    noModels: 'No hay modelos instalados. Descárgalos con «ollama pull».',
+    roleRag: 'Generador RAG', descRoleRag: 'Respuesta final en modo documento',
+    roleChat: 'Subconsultas', descRoleChat: 'Conversación y descomposición de consultas',
+    roleContextual: 'Recuperación contextual', descRoleContextual: 'Enriquece fragmentos al indexar',
+    roleRecomp: 'Síntesis RECOMP', descRoleRecomp: 'Resume el contexto antes de generar',
+    modelSaveError: 'No se pudo cambiar el modelo.',
+  },
+  en: {
+    tabDocs: 'Documents', tabPipeline: 'RAG Pipeline',
+    corpusError: 'Could not switch store.',
+    corpusConflict: 'Indexing in progress; please wait.',
+    corpusConnError: '✗ Connection error while switching corpus.',
+    collection: 'Collection ({n} docs)', noDocs: 'No documents indexed',
+    deleteDoc: 'Delete document', viewPdf: 'View PDF',
+    indexingFile: 'Indexing {file}', reindexingStatus: 'Re-indexing in progress...',
+    section1: '1. Indexing',
+    labelContextual: 'Contextual Retrieval', descContextual: 'Enrich chunks with LLM',
+    section2: '2. Retrieval',
+    labelHybrid: 'Hybrid Search', descHybrid: 'Semantic + BM25',
+    labelQueryDecomp: 'Query Decomposition', descQueryDecomp: 'Sub-queries via auxiliary LLM',
+    labelImageIndex: 'Image indexing', descImageIndex: 'Direct visual retrieval',
+    labelImageDesc: 'Figure descriptions', descImageDesc: 'Vision-written figure text (needs reindex and a vision model)',
+    section3: '3. Ranking & Context',
+    labelReranker: 'Cross-Encoder Reranker', descReranker: 'Precision reordering',
+    labelExpandContext: 'Expand Context', descExpandContext: 'Add adjacent chunks',
+    labelOptimizeContext: 'Optimize Context', descOptimizeContext: 'Clean PDF artifacts',
+    labelRecomp: 'RECOMP Synthesis', descRecomp: 'Synthesize context with LLM',
+    section4: '4. Re-index',
+    reindexHint: 'Adjust settings above and re-index to apply. Optionally add new PDFs.',
+    addPdfs: 'Add PDFs', remove: 'Remove', reindexBtn: 'Re-index',
+    fragments: '{n} fragments', ollamaStatus: 'Ollama Local',
+    clearChat: 'Clear chat', youLabel: 'You',
+    themeLight: 'Light mode', themeDark: 'Dark mode', close: 'Close',
+    sources: '{n} sources', copyMsg: 'Copy message',
+    placeholderRag: 'Ask about your documents…', placeholderChat: 'Type a message…',
+    footerMode: 'Document mode', footerModeChat: 'Conversation mode',
+    indexingTitle: 'Indexing documents',
+    indexingHint: 'This may take a few minutes depending on your hardware.',
+    processing: 'Processing:', fileUnit: 'file', fileUnitPlural: 'files',
+    autoRefresh: 'The page will refresh automatically when done.',
+    retry: 'Retry', connErrorTitle: 'Connection error',
+    connecting: 'Connecting to MonkeyGrab…',
+    reindexError: '✗ Error: {error}', reindexConnError: '✗ Connection error.',
+    confirmDelete: 'Delete "{name}"? It will be removed from the index and disk.',
+    docDeleted: '✓ Document **{name}** deleted.',
+    deleteError: '✗ Error deleting: {error}',
+    deleteConnError: '✗ Connection error while deleting.',
+    unknownError: 'Unknown error', queryError: 'Query error',
+    modelError: 'Error in model response',
+    connClosed: 'The connection closed before the response was complete.',
+    settingsSaveError: 'Could not save setting.',
+    noServer: 'Could not connect to the server. Is Flask running?',
+    retryError: 'Retry failed',
+    indexingFailed: 'Indexing could not be completed.',
+    noResults: 'No relevant information found in the documents.',
+    tabModels: 'Models',
+    modeStudy: 'STUDY',
+    studyKindSummary: 'Summary', studyKindOutline: 'Outline', studyKindQuiz: 'Quiz',
+    studyPlaceholder: 'Type a document name...',
+    studyHint: 'Pick what to generate and type the document. Tab to autocomplete.',
+    studyNoMatch: 'No document matches "{query}".',
+    studyAnswer: 'Answer', studyShowAnswersInline: 'Show answers',
+    studyEmptyTitle: 'Study your documents',
+    studyEmptyBody: 'A summary, an outline or a quiz from any indexed PDF.',
+    studyPages: 'Pages', studyPagesShort: 'pages',
+    studyMalformed: 'The model did not follow the format. Try again, or another model.',
+    studyFailed: 'Could not generate: {error}',
+    storesLabel: 'Vector store',
+    storeNotIndexed: 'not indexed',
+    storeEmptyHint: 'Empty store. Upload PDFs and re-index to activate it.',
+    fingerprintStaleWarning: 'The saved index no longer matches the active configuration, but the app keeps working with it. Re-index whenever you like: it can take an hour or more.',
+    modelsRoles: 'Model roles',
+    ollamaTitle: 'Ollama server',
+    ollamaOnline: 'Running',
+    ollamaOffline: 'Stopped',
+    ollamaStartBtn: 'Start Ollama',
+    ollamaStarting: 'Starting…',
+    ollamaStartFailed: 'Could not start Ollama.',
+    ollamaNotFound: 'Ollama executable not found on PATH.',
+    refreshModels: 'Refresh list',
+    noModels: 'No models installed. Pull some with “ollama pull”.',
+    roleRag: 'RAG generator', descRoleRag: 'Final answer in document mode',
+    roleChat: 'Sub-queries', descRoleChat: 'Conversation and query decomposition',
+    roleContextual: 'Contextual retrieval', descRoleContextual: 'Enriches chunks at indexing',
+    roleRecomp: 'RECOMP synthesis', descRoleRecomp: 'Summarizes context before generation',
+    modelSaveError: 'Could not change the model.',
+  },
+  ca: {
+    tabDocs: 'Documents', tabPipeline: 'Pipeline RAG',
+    corpusError: 'No s\'ha pogut canviar el magatzem.',
+    corpusConflict: 'Hi ha indexació en curs; espera un moment.',
+    corpusConnError: '✗ Error de connexió en canviar el corpus.',
+    collection: "Col·lecció ({n} docs)", noDocs: 'No hi ha documents indexats',
+    deleteDoc: 'Eliminar document', viewPdf: 'Veure PDF',
+    indexingFile: 'Indexant {file}', reindexingStatus: 'Re-indexació en curs...',
+    section1: '1. Indexació',
+    labelContextual: 'Recuperació contextual', descContextual: 'Enriqueix fragments amb LLM',
+    section2: '2. Recuperació',
+    labelHybrid: 'Cerca híbrida', descHybrid: 'Semàntica + BM25',
+    labelQueryDecomp: 'Descomposició de consultes', descQueryDecomp: 'Sub-consultes amb LLM auxiliar',
+    labelImageIndex: 'Indexat d\'imatges', descImageIndex: 'Recuperació visual directa',
+    labelImageDesc: 'Descripció de figures', descImageDesc: 'Text descriptiu per visió (cal reindexar i model amb visió)',
+    section3: '3. Rànquing i context',
+    labelReranker: 'Reordenador cross-encoder', descReranker: 'Reordenament de precisió',
+    labelExpandContext: 'Expandir context', descExpandContext: 'Afig fragments adjacents',
+    labelOptimizeContext: 'Optimitzar context', descOptimizeContext: 'Neteja artefactes PDF',
+    labelRecomp: 'Síntesi RECOMP', descRecomp: 'Sintetitza context amb LLM',
+    section4: '4. Re-indexació',
+    reindexHint: "Ajusta les opcions dalt i torna a indexar per a aplicar-les. Opcionalment afig PDFs nous.",
+    addPdfs: 'Afegir PDFs', remove: 'Llevar', reindexBtn: 'Re-indexar',
+    fragments: '{n} fragments', ollamaStatus: 'Ollama Local',
+    clearChat: 'Netejar xat', youLabel: 'Tu',
+    themeLight: 'Mode clar', themeDark: 'Mode fosc', close: 'Tancar',
+    sources: '{n} fonts', copyMsg: 'Copiar missatge',
+    placeholderRag: 'Pregunta sobre els teus documents…', placeholderChat: 'Escriu un missatge…',
+    footerMode: 'Mode document', footerModeChat: 'Mode conversa',
+    indexingTitle: "S'estan indexant els documents",
+    indexingHint: 'Pot tardar uns minuts depenent del teu maquinari.',
+    processing: 'Processant:', fileUnit: 'arxiu', fileUnitPlural: 'arxius',
+    autoRefresh: "La pàgina s'actualitzarà automàticament en acabar.",
+    retry: 'Reintentar', connErrorTitle: 'Error de connexió',
+    connecting: 'Connectant amb MonkeyGrab…',
+    reindexError: '✗ Error: {error}', reindexConnError: '✗ Error de connexió.',
+    confirmDelete: 'Eliminar "{name}"? S\'esborrarà de l\'índex i del disc.',
+    docDeleted: '✓ Document **{name}** eliminat.',
+    deleteError: '✗ Error en eliminar: {error}',
+    deleteConnError: '✗ Error de connexió en eliminar.',
+    unknownError: 'Error desconegut', queryError: 'Error en la consulta',
+    modelError: 'Error en la resposta del model',
+    connClosed: 'La connexió es va tancar abans de completar la resposta.',
+    settingsSaveError: "No s'ha pogut guardar l'ajust.",
+    noServer: "No s'ha pogut connectar amb el servidor. Està Flask executant-se?",
+    retryError: 'Error en reintentar',
+    indexingFailed: "La indexació no s'ha pogut completar.",
+    noResults: "No s'ha trobat informació rellevant als documents.",
+    tabModels: 'Models',
+    tabStudy: 'Estudiar',
+    modeStudy: 'ESTUDIAR',
+    studyKindSummary: 'Resum', studyKindOutline: 'Esquema', studyKindQuiz: 'Qüestionari',
+    studyPlaceholder: 'Escriu el nom d\'un document...',
+    studyHint: 'Tria què generar i escriu el document. Tab per autocompletar.',
+    studyNoMatch: 'No trobe cap document que coincidisca amb «{query}».',
+    studyAnswer: 'Resposta', studyShowAnswersInline: 'Veure respostes',
+    studyEmptyTitle: 'Estudia els teus documents',
+    studyEmptyBody: 'Un resum, un esquema o un qüestionari de qualsevol PDF indexat.',
+    studyPages: 'Pàgines', studyPagesShort: 'pàgines',
+    studyMalformed: 'El model no ha respectat el format. Prova una altra vegada, o amb un altre model.',
+    studyFailed: 'No s\'ha pogut generar: {error}',
+    storesLabel: 'Magatzem vectorial',
+    storeNotIndexed: 'sense indexar',
+    storeEmptyHint: 'Magatzem buit. Puja PDFs i reindexa per a activar-lo.',
+    fingerprintStaleWarning: "L'índex guardat no coincideix amb la configuració activa, però l'app continua funcionant amb ell. Reindexa quan vulgues: pot trigar una hora o més.",
+    modelsRoles: 'Rols de model',
+    ollamaTitle: 'Servidor Ollama',
+    ollamaOnline: 'En execució',
+    ollamaOffline: 'Aturat',
+    ollamaStartBtn: 'Arrancar Ollama',
+    ollamaStarting: 'Arrancant…',
+    ollamaStartFailed: "No s'ha pogut arrancar Ollama.",
+    ollamaNotFound: "No s'ha trobat l'executable d'Ollama al PATH.",
+    refreshModels: 'Actualitzar llista',
+    noModels: 'No hi ha models instal·lats. Descarrega\'n amb «ollama pull».',
+    roleRag: 'Generador RAG', descRoleRag: 'Resposta final en mode document',
+    roleChat: 'Subconsultes', descRoleChat: 'Conversa i descomposició de consultes',
+    roleContextual: 'Recuperació contextual', descRoleContextual: 'Enriqueix fragments en indexar',
+    roleRecomp: 'Síntesi RECOMP', descRoleRecomp: 'Resumeix el context abans de generar',
+    modelSaveError: 'No s\'ha pogut canviar el model.',
+  },
+} as const;
+
+export const LANG_OPTIONS: Array<{ code: Lang; label: string }> = [
+  { code: 'es', label: 'ES' },
+  { code: 'en', label: 'EN' },
+  { code: 'ca', label: 'VAL' },
+];
+
+export function normalizeLang(value: string | null): Lang {
+  return value === 'en' || value === 'ca' || value === 'es' ? value : 'es';
+}
+
+export function fill(tpl: string, vars: Record<string, string | number>): string {
+  return tpl.replace(/\{(\w+)\}/g, (_, k) => String(vars[k] ?? ''));
+}
+
+/** The table for one language.
+ *
+ * Components take this instead of ``Record<string, string>`` so a mistyped key
+ * is a compile error rather than an ``undefined`` rendered into the page.
+ */
+export type Strings = (typeof STRINGS)[Lang];
